% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRLineTypes.R
\name{generateRLineTypes}
\alias{generateRLineTypes}
\title{Generate a plot of line types which R knows about.}
\usage{
generateRLineTypes()
}
\description{
Generate line types available in R.
}
\examples{

#The code to generate line types available in R
generateRLineTypes()

}
\author{
Alboukadel Kassambara <alboukadel.kassambara@gmail.com>
}
\references{
http://www.sthda.com
}
\seealso{
\code{\link{generateRPointShapes}}
}

