% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRPointShapes.R
\name{generateRPointShapes}
\alias{generateRPointShapes}
\title{Generate a plot of point shapes which R knows about.}
\usage{
generateRPointShapes()
}
\description{
Generate point shapes available in R
}
\examples{

#The code to generate point shapes available in R
generateRPointShapes()

}
\author{
Alboukadel Kassambara <alboukadel.kassambara@gmail.com>
}
\references{
http://www.sthda.com
}
\seealso{
\code{\link{generateRLineTypes}}
}

