% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.barplot.R
\name{ggplot2.barplot}
\alias{ggplot2.barplot}
\title{Easy barplot plot with R package ggplot2}
\usage{
ggplot2.barplot(data, xName = NULL, yName = NULL, groupName = NULL,
  groupColors = NULL, brewerPalette = NULL, ...)
}
\arguments{
\item{data}{data.frame or a numeric vector. Columns are variables and rows
are observations.}

\item{xName}{The name of column containing x variable (i.e groups). Default
value is NULL.}

\item{yName}{The name of column containing y variable. If yName=NULL, data
should be a numeric vector.}

\item{groupName}{The name of column containing group variable. This variable
is used to color plot according to the group.}

\item{groupColors}{Color of groups. groupColors should have the same length
as groups.}

\item{brewerPalette}{This can be also used to indicate group colors. In this
case the parameter groupColors should be NULL. e.g: brewerPalette="Paired".}

\item{\dots}{Other arguments passed on to ggplot2.customize custom function
or to geom_bar functions from ggplot2 package.}
}
\value{
a ggplot
}
\description{
Easy barplot plot with R package ggplot2
}
\examples{

#data
df <- data.frame(time = factor(c("Lunch","Dinner"), levels=c("Lunch","Dinner")),
                 total_bill = c(14.89, 17.23))

#plot
ggplot2.barplot(data=df, xName="time", yName='total_bill',
                mainTitle="Plot of total bill\\n per time of day",
                xtitle="Time of day", ytitle="Total bill")

#Or use this
plot<-ggplot2.barplot(data=df, xName="time", yName='total_bill')
plot<-ggplot2.customize(plot,  mainTitle="Plot of total bill\\n per time of day",
                      xtitle="Time of day", ytitle="Total bill")
print(plot)

}
\author{
Alboukadel Kassambara <alboukadel.kassambara@gmail.com>
}
\references{
http://www.sthda.com
}
\seealso{
\code{\link{ggplot2.lineplot}, \link{ggplot2.histogram}}
}

