% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.density.R
\name{ggplot2.density}
\alias{ggplot2.density}
\title{Easy density plot with R package ggplot2}
\usage{
ggplot2.density(data, xName = NULL, groupName = NULL, addMeanLine = FALSE,
  meanLineColor = NULL, meanLineType = "dashed", meanLineSize = 1,
  densityFill = NULL, fillGroupDensity = FALSE,
  colorGroupDensityLine = FALSE, groupColors = NULL, brewerPalette = NULL,
  ...)
}
\arguments{
\item{data}{data.frame or a numeric vector. Columns are variables and rows
are observations.}

\item{xName}{The name of column containing x variable (i.e groups). Default
value is NULL.}

\item{groupName}{The name of column containing group variable. This variable
is used to color plot according to the group.}

\item{addMeanLine}{if TRUE, the mean line is added on the plot for each
group. Default value is FALSE.}

\item{meanLineColor, meanLineType, meanLineSize}{mean line color, type and
size.}

\item{densityFill}{Fill color of density plot. This is only considered when
groupName=NULL.}

\item{fillGroupDensity}{If TRUE, density curve of each group is filled.
Default value is FALSE.}

\item{colorGroupDensityLine}{If TRUE, density curve line are colored.
Default value is FALSE.}

\item{groupColors}{Color of groups. groupColors should have the same length
as groups.}

\item{brewerPalette}{This can be also used to indicate group colors. In this
case the parameter groupColors should be NULL. e.g: brewerPalette="Paired".}

\item{\dots}{Other arguments passed on to ggplot2.customize custom function
or to geom_density functions from ggplot2 package.}
}
\value{
a ggplot
}
\description{
Easy density plot with R package ggplot2
}
\examples{

#data
set.seed(1234)
df <- data.frame(grp = factor( rep(c("A","B"), each=200) ),
                   var = c(rnorm(200),rnorm(200, mean=.8)))

#plot
ggplot2.density(data=df, xName='var',groupName='grp', groupColors=c('#999999','#E69F00'),
                mainTitle="Plot of variable histogram \\nper group",
                xtitle="Variable", ytitle="Density")

#Or use this
plot<-ggplot2.density(data=df, xName='var',groupName='grp', groupColors=c('#999999','#E69F00'))
plot<-ggplot2.customize(plot, mainTitle="Plot of variable histogram \\nper group",
                        xtitle="Variable", ytitle="Density")
print(plot)

}
\author{
Alboukadel Kassambara <alboukadel.kassambara@gmail.com>
}
\references{
http://www.sthda.com
}
\seealso{
\code{\link{ggplot2.dotplot}, \link{ggplot2.violinplot},
\link{ggplot2.stripchart}, \link{ggplot2.boxplot}, \link{ggplot2.histogram}}
}
\keyword{~kwd1}
\keyword{~kwd2}

