% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.dotplot.R
\name{ggplot2.dotplot}
\alias{ggplot2.dotplot}
\title{Easy dotplot plot with R package ggplot2}
\usage{
ggplot2.dotplot(data, xName = NULL, yName = NULL, groupName = NULL,
  position = position_dodge(0.8), addMean = FALSE, meanPointShape = 5,
  meanPointSize = 4, meanPointColor = "black", meanPointFill = "blue",
  addBoxplot = FALSE, boxplotFill = "white", boxplotColor = "black",
  boxplotLineWeight = 0.5, groupColors = NULL, brewerPalette = NULL, ...)
}
\arguments{
\item{data}{data.frame or a numeric vector. Columns are variables and rows
are observations.}

\item{xName}{The name of column containing x variable (i.e groups). Default
value is NULL.}

\item{yName}{The name of column containing y variable. If yName=NULL, data
should be a numeric vector.}

\item{groupName}{The name of column containing group variable. This variable
is used to color plot according to the group.}

\item{position}{The position adjustment to use for overlappling points.
Default value is position_dodge(0.8).}

\item{addMean}{if TRUE, the mean point is added on the plot for each group.
Default value is FALSE.}

\item{meanPointShape}{The shape of mean point.}

\item{meanPointSize}{The size of mean point}

\item{meanPointColor}{Border color of the mean point. Default value is
"black".}

\item{meanPointFill}{Fill color of mean point. This parameter is used only
when meanPointShape=21 to 25. Default value is "blue"}

\item{addBoxplot}{If TRUE, boxplot is added on the dotplot. Default value is
FALSE.}

\item{boxplotFill}{Fill color of the boxplot. Default value is white.}

\item{boxplotColor}{Boxplot line color. Default value is black.}

\item{boxplotLineWeight}{Boxplot line weight. Default value is 0.5.}

\item{groupColors}{Color of groups. groupColors should have the same length
as groups.}

\item{brewerPalette}{This can be also used to indicate group colors. In this
case the parameter groupColors should be NULL. e.g: brewerPalette="Paired".}

\item{\dots}{Other arguments passed on to ggplot2.customize custom function
or to geom_dotplot functions from ggplot2 package.}
}
\value{
a ggplot
}
\description{
Easy dotplot plot with R package ggplot2
}
\examples{

df <- ToothGrowth
ggplot2.dotplot(data=df, xName='dose',yName='len',
                mainTitle="Plot of length according\\n to the dose",
                xtitle="Dose (mg)", ytitle="Length")

#Or use this
plot<-ggplot2.dotplot(data=df, xName='dose',yName='len')
plot<-ggplot2.customize(plot, mainTitle="Plot of length according\\n to the dose",
                        xtitle="Dose (mg)", ytitle="Length")
print(plot)

}
\author{
Alboukadel Kassambara <alboukadel.kassambara@gmail.com>
}
\references{
http://www.sthda.com
}
\seealso{
\code{\link{ggplot2.boxplot}, \link{ggplot2.violinplot},
\link{ggplot2.stripchart}, \link{ggplot2.density}, \link{ggplot2.histogram},
\link{ggplot2.customize}}
}

