% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.histogram.R
\name{ggplot2.histogram}
\alias{ggplot2.histogram}
\title{Easy histogram plot with R package ggplot2}
\usage{
ggplot2.histogram(data, xName = NULL, groupName = NULL,
  position = c("identity", "stack", "dodge"), addMeanLine = FALSE,
  meanLineColor = NULL, meanLineType = "dashed", meanLineSize = 1,
  addDensityCurve = FALSE, densityFill = "#FF6666", densityAlpha = 0.2,
  densityLineType = "solid", densityLineColor = "#2F2F2F",
  scale = c("frequency", "density"), groupColors = NULL,
  brewerPalette = NULL, fill = "black", color = "black",
  linetype = "solid", ...)
}
\arguments{
\item{data}{data.frame or a numeric vector. Columns are variables and rows
are observations.}

\item{xName}{The name of column containing x variable (i.e groups). Default
value is NULL.}

\item{groupName}{The name of column containing group variable. This variable
is used to color plot according to the group.}

\item{position}{Change the position adjustment to use for overlapping points
on the layer. Possible values for the argument position is "identity",
"stack", "dodge". Default value is identity.}

\item{addMeanLine}{if TRUE, the mean line is added on the plot for each
group. Default value is FALSE.}

\item{meanLineColor, meanLineType, meanLineSize}{mean line color, type and
size.}

\item{addDensityCurve}{If true, add density curve. Default value is FALSE.}

\item{densityFill}{The fill color of density plot. The value is considered
only when groupName=NULL. If groupName is specified, density curves are
colored according groupColors pr brewerPalette.}

\item{densityAlpha}{Degre of transparency of overlaid colors for density
curves. Default is 0.2 (20\%).}

\item{densityLineType, densityLineColor}{Line type and color for density
curve.}

\item{scale}{Indicate whether y axis values are density or frequency.
Default value is frequency.}

\item{groupColors}{Color of groups. groupColors should have the same length
as groups.}

\item{brewerPalette}{This can be also used to indicate group colors. In this
case the parameter groupColors should be NULL. e.g: brewerPalette="Paired".}

\item{linetype, fill, color}{linetype, outline and fill colors}

\item{\dots}{Other arguments passed on to ggplot2.customize custom function
or to geom_histogram and geom_density functions from ggplot2 package.}
}
\value{
a ggplot
}
\description{
Easy histogram plot with R package ggplot2
}
\examples{

#data
set.seed(1234)
df <- data.frame(grp = factor( rep(c("A","B"), each=200) ), 
                   var = c(rnorm(200),rnorm(200, mean=.8)))
#plot
plot<-ggplot2.histogram(data=df, xName='var',groupName='grp', groupColors=c('#999999','#E69F00'))
plot<-ggplot2.customize(plot, mainTitle="Plot of variable histogram \\nper group",
                        xtitle="Variable", ytitle="Histogram")
print(plot)

}
\author{
Alboukadel Kassambara <alboukadel.kassambara@gmail.com>
}
\references{
http://www.sthda.com
}
\seealso{
\code{\link{ggplot2.dotplot}, \link{ggplot2.violinplot},
\link{ggplot2.stripchart}, \link{ggplot2.boxplot}, \link{ggplot2.density}}
}
\keyword{~kwd1}
\keyword{~kwd2}

