% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.multiplot.R
\name{ggplot2.multiplot}
\alias{ggplot2.multiplot}
\title{Put multiple graphs on the same page.}
\usage{
ggplot2.multiplot(..., plotlist = NULL, cols = 2)
}
\arguments{
\item{plotlist}{a list of plots}

\item{cols}{Number of columns in layout}

\item{\dots}{List of ggplot2 objects separated by a comma. (e.g: plot1,
plot2, plot3)}
}
\value{
a ggplot
}
\description{
Put multiple graphs on the same page.
}
\examples{

#data.frame
df <- ToothGrowth
#Custom boxplot with centered dot plot
plot1<-ggplot2.boxplot(data=df, xName='dose',yName='len', groupName='dose',
                addDot=TRUE, dotSize=1)

#Custom dotplot with centered dot plot
plot2<-ggplot2.dotplot(data=df, xName='dose',yName='len', groupName='dose')

#Custom stripchart with centered dot plot
plot3<-ggplot2.stripchart(data=df, xName='dose',yName='len', groupName='dose')

#Notched boxplot
plot4<-ggplot2.boxplot(data=df, xName='dose',yName='len',
                notch=TRUE)

#Multple graphs on the same page
ggplot2.multiplot(plot1,plot2,plot3,plot4, cols=2)

}
\author{
Alboukadel Kassambara <alboukadel.kassambara@gmail.com>
}
\references{
http://www.sthda.com
}

