% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.setAxis.R
\name{ggplot2.setAxis}
\alias{ggplot2.setAxis}
\title{Customize x and y axis of a plot generated with ggplot2 package}
\usage{
ggplot2.setAxis(plot, xShowTitle = TRUE, yShowTitle = TRUE, xtitle = NULL,
  ytitle = NULL, xtitleFont = c(14, "bold", "black"), ytitleFont = c(14,
  "bold", "black"), xlim = NULL, ylim = NULL, xScale = c("none", "log2",
  "log10"), yScale = c("none", "log2", "log10"), xShowTickLabel = TRUE,
  yShowTickLabel = TRUE, xTickLabelFont = c(12, "bold", "black"),
  yTickLabelFont = c(12, "bold", "black"), xtickLabelRotation = 0,
  ytickLabelRotation = 0, hideAxisTicks = FALSE, axisLine = c(0.5,
  "solid", "#E5E5E5"))
}
\arguments{
\item{plot}{A plot object generated with ggplot2 or easyGgplot2 packages
(boxplot, densityplot, histogram, dotplot, ....)}

\item{xShowTitle, yShowTitle}{if TRUE, x axis and y axis titles will be
shown. \strong{Default values are TRUE.}}

\item{xtitle, ytitle}{x and y axis labels.}

\item{xtitleFont, ytitleFont}{A vector of length 3 indicating respectively
the size, the style and the color of x and y axis titles. \strong{Default
values are c(14,"bold", "black").}}

\item{xlim, ylim}{Limit for the x and y axis. \strong{Default values are
NULL}.}

\item{xScale, yScale}{x and y axis scales. Possible values are c("none",
"log2", "log10"). Example: yScale="log2". \strong{Default values are
"none".}}

\item{xShowTickLabel, yShowTickLabel}{if TRUE, x and y axis tick mark labels
will be shown. \strong{Default values are TRUE.}}

\item{xTickLabelFont, yTickLabelFont}{A vector of length 3 indicating
respectively the size, the style and the color of x and y axis tick label
fonts. \strong{Default values are c(12, "bold", "black").}}

\item{xtickLabelRotation, ytickLabelRotation}{Rotation angle of x and y axis
tick labels. \strong{Default values are 0.}}

\item{hideAxisTicks}{if TRUE, x and y axis ticks are removed.
\strong{Default value is FALSE}.}

\item{axisLine}{A vector of length 3 indicating respectively the size, the
line type and the color of axis lines.\strong{ Default value is c(0.5,
"solid", "#E5E5E5").}}
}
\value{
a ggplot
}
\description{
Customize x and y axis of a plot generated with ggplot2 package
}
\examples{

#Boxplot using easyGgplot2 package
p<-ggplot2.boxplot(data=ToothGrowth, xName='dose',yName='len')
ggplot2.setAxis(p, xtitle="Dose (mg)", ytitle="length", 
                axisLine=c(0.5, "solid", "black"))

}
\author{
Alboukadel Kassambara <alboukadel.kassambara@gmail.com>
}
\references{
http://www.sthda.com
}
\seealso{
\code{\link{ggplot2.customize}}
}

