% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rquery.loadPackages.R
\name{rquery.loadPackages}
\alias{rquery.loadPackages}
\title{Installation and loading of packages}
\usage{
rquery.loadPackages(pkgs)
}
\arguments{
\item{pkgs}{The packages to load.}
}
\description{
Missing packages are automatically installed and loaded.
}
\examples{

rquery.loadPackages(c("ggplot2", "plyr"))

}
\author{
Alboukadel Kassambara <alboukadel.kassambara@gmail.com>
}
\references{
http://www.sthda.com
}
\keyword{~kwd1}
\keyword{~kwd2}

