% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showCols.R
\name{showCols}
\alias{showCols}
\title{Generate a plot of color names which R knows about.}
\usage{
showCols(cl = colors(), bg = "grey", cex = 0.75, rot = 30)
}
\arguments{
\item{cl}{A vector of colors to plot. Default value is the list of color
names returned by R color() function.}

\item{bg}{Background color of the plot}

\item{cex}{The size of the text.}

\item{rot}{The rotation angle of the text}
}
\value{
a ggplot
}
\description{
Generate R color names.
}
\examples{

col=colors() #list of color names
showCols(col[1:100]) #plot the first 100 colors

}
\author{
Alboukadel Kassambara <alboukadel.kassambara@gmail.com>
}
\references{
http://www.sthda.com
}
\seealso{
\code{\link{generateRLineTypes}, \link{generateRPointShapes}}
}

